Fix unique title validator
